Moving code to its own package. Make data loading location independent.
