+++ .gitignore
@@ -1,3 +1,4 @@
+*.pyc
 resources/
 resources.tar.gz
 __pycache__/
