+++ StanceClassifier/features/heuristics.py
