+++ StanceClassifier/server/Run_HTTP_StanceClassifier_Server.py
@@ -4,7 +4,7 @@ from struct import pack
 import sys
 import socket, json
 sys.path[0:0] = ["util/"]
-from util import Util
+from StanceClassifier.util import Util
 
 class StanceClassifier:
